Refactor overflow menu to use id.xml resources
