+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/family/profile/FamilyProfileEvent.kt
@@ -30,7 +30,7 @@ sealed class FamilyProfileEvent {
   data class OpenMemberProfile(val patientId: String, val navController: NavHostController) :
     FamilyProfileEvent()
 
-  data class OverflowMenuClick(val menuId: String) : FamilyProfileEvent()
+  data class OverflowMenuClick(val menuId: Int) : FamilyProfileEvent()
 
   data class FetchFamilyProfileData(val familyHeadId: String?) : FamilyProfileEvent()
