+++ android/quest/src/main/res/values/strings.xml
@@ -27,5 +27,12 @@
     <string name="add_memeber">ADD MEMBER</string>
     <string name="what_to_do">What do you want to do?</string>
     <string name="view_profile">View profile</string>
+    <string name="family_details">Family details</string>
+    <string name="change_family_head">Change family head</string>
+    <string name="change_primary_caregiver">Change primary caregiver</string>
+    <string name="family_activity">Family activity</string>
+    <string name="view_past_encounters">View past encounters</string>
+    <string name="remove_family">Remove family</string>
+    <string name="family_suffix">%1$s Family</string>
 
 </resources>
